traits: Rename bitcoin to bitcoin_rpc.
