+++ core/src/traits/bitcoin_rpc.rs
